refactor : 패키지 분리, 테스트 분리
