--- src/main/java/study/step2/LottoNumberGenerator.java
@@ -1,8 +1,0 @@
-package study.step2;
-
-import java.util.List;
-
-@FunctionalInterface
-public interface LottoNumberGenerator {
-    List<Integer> getLottoNumbers();
-}
